requirements.txt
Django==1.11.2
django-tables2==1.21.1
djangorestframework==3.7.7
honcho==1.0.1
pytz==2017.2
uWSGI==2.0.15
